+++ models/room_message.go
@@ -6,8 +6,8 @@ import (
 
 type RoomMessage struct {
 	gorm.Model
-	UserID    *uint64            `gorm:""`
-	RoomID    *uint64            `gorm:""`
+	UserID    *uint              `gorm:""`
+	RoomID    *uint              `gorm:""`
 	Type      string             `gorm:"size:255;not null"`
 	Body      string             `gorm:"type:text;not null"`
 	Reactions []*ContentReaction `gorm:"foreignKey:MessageID"`
